requirements.txt
# Django Framework
django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0

# Supabase Client
supabase==2.22.2

# AI Integration
groq>=0.32.0

# Utilities
python-dotenv==1.0.0
